optimize performance

+++ Coosu.Storyboard/Animation.cs
@@ -11,7 +11,7 @@ namespace Coosu.Storyboard
 
 
         public int FrameCount { get; set; }
-        public double FrameDelay { get; set; }
+        public float FrameDelay { get; set; }
         public LoopType LoopType { get; set; }
 
         /// <summary>
@@ -25,8 +25,8 @@ namespace Coosu.Storyboard
         /// <param name="frameCount">Set frame count.</param>
         /// <param name="frameDelay">Set frame rate (frame delay).</param>
         /// <param name="loopType">Set loop type.</param>
-        public Animation(LayerType layerType, OriginType originType, string imagePath, double defaultX,
-            double defaultY, int frameCount, double frameDelay, LoopType loopType)
+        public Animation(LayerType layerType, OriginType originType, string imagePath, float defaultX,
+            float defaultY, int frameCount, float frameDelay, LoopType loopType)
             : base(layerType, originType, imagePath, defaultX, defaultY)
         {
             FrameCount = frameCount;
@@ -48,8 +48,8 @@ namespace Coosu.Storyboard
         public Animation(ReadOnlySpan<char> layer, 
             ReadOnlySpan<char> origin,
             ReadOnlySpan<char> imagePath,
-            double defaultX, double defaultY, 
-            int frameCount, double frameDelay, ReadOnlySpan<char> loopType)
+            float defaultX, float defaultY, 
+            int frameCount, float frameDelay, ReadOnlySpan<char> loopType)
             : base(layer, origin, imagePath, defaultX, defaultY)
         {
             FrameCount = frameCount;
