+++ Coosu.Storyboard/Common/ISpriteHost.cs
@@ -5,10 +5,10 @@ namespace Coosu.Storyboard.Common
 {
     public interface ISpriteHost : IScriptable, ICloneable, IEnumerable<Sprite>
     {
-        double MaxTime { get; }
-        double MinTime { get; }
-        double MaxStartTime { get; }
-        double MinEndTime { get; }
+        float MaxTime { get; }
+        float MinTime { get; }
+        float MaxStartTime { get; }
+        float MinEndTime { get; }
         IList<Sprite> Sprites { get; }
         IList<ISpriteHost> SubHosts { get; }
         Camera2 Camera2 { get; }
