+++ Coosu.Storyboard/Extensibility/EventCreationDelegate.cs
@@ -5,5 +5,5 @@ using Coosu.Storyboard.Events;
 namespace Coosu.Storyboard.Extensibility
 {
     public delegate BasicEvent EventCreationDelegate(EventType e, EasingFunctionBase easing,
-        double startTime, double endTime, List<double> values);
+        float startTime, float endTime, List<float> values);
 }
